Move cargo crate vendoring into its own module
